+++ samples/Samples.Mvc5/Global.asax.cs
@@ -112,6 +112,7 @@ namespace Samples.Mvc5
                 // specified position in the .RenderIncludes() call.
                 PopupRenderPosition = RenderPosition.Right,  // defaults to left
                 PopupMaxTracesToShow = 10,                   // defaults to 15
+                PopupDecimalPlaces = 1,                      // defaults to 2
                 ColorScheme = ColorScheme.Auto,              // defaults to light
 
                 // ResultsAuthorize (optional - open to all by default):